requirements.txt
asgiref==3.5.2
Django==4.1.2
django-ckeditor==6.5.1
django-environ==0.9.0
django-js-asset==2.0.0
mysqlclient==2.1.1
Pillow==9.2.0
python-decouple==3.6
sqlparse==0.4.3
